backend/requirements.txt
uvicorn
fastapi
python-osc
soundfile
torch
httpx
pyfiglet
Pillow
TTS